pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Computer Vision"
version = "0.0.1"
description = ""
authors = [
    "Roger Waldvogel <roger.waldvogel@stud.hslu.ch>",
    "Nicco Vöggeli <nicco.voegeli@stud.hslu.ch>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5"
numpy = "^1.24.1"
tensorflow = "2.11.0"
setuptools = "^67.0.0"
tensorboard = "^2.11.2"
Pillow = "9.4.0"
scikit-learn= "1.2.1"
matplotlib = "3.6.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.16.1"
